Clean up JNI copying: extract to separate file, use idiomatic val injection, assert non-null outputDirectory
